Updated instructions since QCMA ignores the xml file by default
By default new releases of QCMA default to ignore the `psp2-updatelist.xml` file.  QCMA also defaults the `Use this version for updates` to 3.60, which means that anyone trying to update from 3.60-3.64 will be told they are already on the latest version.

So I updated the steps to tell QCMA that the latest version it should allow is 3.65.